backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
pymodbus==3.8.3
paho-mqtt==2.1.0
pyyaml==6.0.2
tinytuya==1.16.1
zeroconf==0.136.2
